1. Added the ability to chart the mouse spring length, force magnitude and damping component magnitude. 2. Got the Firmata controller running to the point where I can define digital outputs in the GUI and control them on an Arduino board. 3. Added a SWIG interface for StdUtils. This is just a test start on getting SWIG setup for all libraries. I tried to add a python SWIG script as well as a .net one, but the python one was not wanting to compile. It will require some more work.

+++ include/StdADT.h
@@ -630,9 +630,9 @@ public:
 
     bool Equal(const CStdPoint<T> &oPoint, double fltTolerance)
     {
-        if( (fabs(x - oPoint.x) < fltTolerance) &&
-            (fabs(y - oPoint.y) < fltTolerance) &&
-            (fabs(z - oPoint.z) < fltTolerance) )
+        if( (fabs((double)(x - oPoint.x)) < fltTolerance) &&
+            (fabs((double)(y - oPoint.y)) < fltTolerance) &&
+            (fabs((double)(z - oPoint.z)) < fltTolerance) )
             return true;
         else
             return false;
@@ -803,7 +803,7 @@ public:
 	};
 
 	double Magnitude()
-	{return sqrt( (x*x) + (y*y) + (z*z) );};
+	{return sqrt((double)((x*x) + (y*y) + (z*z)) );};
 
 	void Normalize()
 	{
@@ -827,11 +827,11 @@ public:
 	//If it is then it just sets it to zero.
 	void ClearNearZero(float fltTolerance = 1e-5f)
 	{
-		if(fabs(x) < fltTolerance)
+		if(fabs((float)x) < fltTolerance)
 			x = 0;
-		if(fabs(y) < fltTolerance)
+		if(fabs((float)y) < fltTolerance)
 			y = 0;
-		if(fabs(z) < fltTolerance)
+		if(fabs((float)z) < fltTolerance)
 			z = 0;
 	}
